pyproject.toml
[project]
name = "qliqy-webapi"
version = "0.1.0"
description = "Qliqy | WebAPI Services"
authors = [{ name = "Ernestor1337", email = "berezin.505@gmail.com" },{ name = "ilia2003", email= "fedorenkoilia@gmail.com"},{ name = "Jeishod", email = "jeishod@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.17.1",
    "asyncpg==0.30.0",
    "fastapi[all]==0.121.0",
    "faststream[rabbit]==0.6.3",
    "greenlet==3.2.4",
    "gunicorn==23.0.0",
    "httpx==0.28.1",
    "loguru==0.7.3",
    "passlib==1.7.4",
    "poethepoet==0.37.0",
    "psycopg2-binary==2.9.11",
    "python-jose==3.5.0",
    "sqlalchemy==2.0.44",
    "uvicorn==0.38.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "faker>=40.1.2",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff==0.14.3",
    "tzdata>=2025.3",
]


[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
# ignore = ["E501", "PLR0911", "PLR0912", "PLR0915", "PLR0913", "PLW2901"]
ignore = [
    "PLR0913",  # too many positional arguments for method call
    "PLW2901",  # overwritten argument in 'for' loop
]
select = [
    # flake8-bugbear
    "B",
    # pycodestyle
    "E",
    "W",
    "FAST",
    # pylint
    "PL",
    # pyflakes
    "F",
    # isort
    "I",
    # pydocstyle
#    "D",
]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.format]
line-ending = "lf"
docstring-code-format = true

[tool.poe]
envfile = [".env"]
verbosity = 1

[tool.poe.executor]
type = "virtualenv"
location = ".venv"

# formatting
[tool.poe.tasks.ruff-lint]
shell = "uv run ruff check --fix app"
[tool.poe.tasks.ruff-format]
shell = "uv run ruff format app"
[tool.poe.tasks.format]
sequence = [
    {ref = "ruff-lint"},
    {ref = "ruff-format"},
]

# migrations
[tool.poe.tasks.create-migration]
cmd = "alembic revision --autogenerate -m \"${message}\""
args = { message = { options = ["-m"], default = "autogenerated_migration" }}
[tool.poe.tasks.apply-migrations]
cmd = "alembic upgrade head"
[tool.poe.tasks.revert-migration]
cmd = "alembic downgrade -1"

# uploading fixtures
[tool.poe.tasks.init-db]
script = "dev.init_database:main_init()"
[tool.poe.tasks.seed-db]
script = "dev.init_database:main_seed()"
[tool.poe.tasks.drop-db]
script = "dev.init_database:main_drop()"
[tool.poe.tasks.load-fixtures]
script = "dev.load_fixtures:main()"
[tool.poe.tasks.load-qa]
script = "dev.load_qa_bases:main()"
[tool.poe.tasks.load-qa-force]
script = "dev.load_qa_bases:main(force=True)"

# starting dev service
[tool.poe.tasks.dev]
cmd = "uvicorn --reload --use-colors --host 0.0.0.0 --port \"${port}\" --log-level debug 'app.main:app'"
args = { port = { options = ["-p"], default = "8000" }}
[tool.poe.tasks.start-dev]
sequence = [
    {ref = "apply-migrations"},
    {ref = "dev"}
]

# starting prod service
[tool.poe.tasks.prod]
cmd = "gunicorn -c app/gunicorn_conf.py app.main:app"
[tool.poe.tasks.start-prod]
sequence = [
    {ref = "apply-migrations"},
    {ref = "prod"}
]

# tests
[tool.poe.tasks.test]
shell = "uv run pytest -q"

[tool.poe.tasks.test-verbose]
shell = "uv run pytest -vv"

[tool.poe.tasks.test-user-forms]
shell = "uv run pytest -q dev/tests/db/user_forms"
